Hiding force rewrite title option when `wp_get_document_title` (thus WordPress 4.4) is available.

+++ admin/pages/metas.php
@@ -34,9 +34,11 @@ $yform->admin_header( true, 'wpseo_titles' );
 					</th>
 					<td>
 						<?php
-						$yform->checkbox( 'forcerewritetitle', __( 'Enable force rewrite titles', 'wordpress-seo' ) );
-						/* translators: %1$s expands to Yoast SEO */
-						echo '<p class="description">', sprintf( __( '%1$s has auto-detected whether it needs to force rewrite the titles for your pages, if you think it\'s wrong and you know what you\'re doing, you can change the setting here.', 'wordpress-seo' ), 'Yoast SEO' ) . '</p>';
+						if ( ! function_exists( 'wp_get_document_title' ) ) {
+							$yform->checkbox( 'forcerewritetitle', __( 'Enable force rewrite titles', 'wordpress-seo' ) );
+							/* translators: %1$s expands to Yoast SEO */
+							echo '<p class="description">', sprintf( __( '%1$s has auto-detected whether it needs to force rewrite the titles for your pages, if you think it\'s wrong and you know what you\'re doing, you can change the setting here.', 'wordpress-seo' ), 'Yoast SEO' ) . '</p>';
+						}
 						?>
 					</td>
 				</tr>
